requirements.txt
fastapi
uvicorn
langchain-core
langchain-groq
langchain-text-splitters
pypdf
python-docx
openpyxl
python-multipart
requests